Made layer component
