+++ src/components/burger-constructor/burger-constructor.module.css
@@ -23,33 +23,6 @@
     overflow-y: scroll;
 }
 
-.product {
-    padding-bottom: 16px;
-}
-.product:last-of-type {
-    padding-bottom: 0;
-}
-
-.layer_element {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    position: relative;
-}
-
-.delete_button {
-    background-color: rgb(240, 240, 240, 0);
-    border: none;
-    padding: 0;
-    height: 24px;
-    align-items: center;
-    cursor: pointer;
-    position: absolute;
-    right: 30px;
-    
-}
-
 .order_box {
     display: flex;
     flex-direction: row;
