+++ src/components/ingredient/ingredient.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState } from "react";
 import PropTypes from "prop-types";
 import styles from './ingredient.module.css';
 import { CurrencyIcon, Counter } from '@ya.praktikum/react-developer-burger-ui-components';
@@ -7,25 +7,25 @@ import { useSelector } from "react-redux";
 
 
 export function Ingredient({ card, onClick }) {
-    const [ { opacity, isDragging } ,dragRef] = useDrag({
+    const [ { opacity } ,dragRef] = useDrag({
         type: 'item', /*card.type*/
         item: card,
         collect: monitor => ({
             opacity: monitor.isDragging() ? 0.5 : 1,
-            isDragging: monitor.isDragging()
         })
     })
     const counter = useSelector(state => state.counter[card._id]);
+    
     return(
         //представление в разделе ингредиентов, надо дополнить отображение для конструктора еще
-      <div className={styles.ingredient_available} key={card._id} onClick={onClick} ref={dragRef} style={ {opacity} } >
+      <div className={styles.ingredient_available} key={card._id} onClick={onClick} ref={dragRef} style={ {opacity} } draggable >
             <img src={card.image} alt={card.name} />
             <p className="text text_type_main-default">{card.name}</p>
             <div className={styles.price_container}>
                 <p className="text text_type_main-default pr-1">{card.price}</p>
                 <CurrencyIcon />
             </div>
-            { counter ? <Counter count={counter} size="default" />  :  <div count={counter}/> }
+            { counter ? <Counter count={counter} size="default" />  :  <div /> }
       </div>      
     )  
 }
