+++ src/components/layer/layer.module.css
@@ -1,0 +1,7 @@
+.layer_element {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    position: relative;
+}
